basic project refactor
